+++ src/test/java/Algo3TP2/EntidadesTests/UnidadMovibleMovimientosTest.java
@@ -1,4 +1,4 @@
-package EntidadesTests;
+package Algo3TP2.EntidadesTests;
 
 import Algo3TP2.Modelos.Bando;
 import Algo3TP2.Modelos.Casillero.Casillero;
